+++ src/components/UI/Modals/index.js
@@ -1,4 +1,4 @@
 import EditProfileModal from './EditProfileModal/EditProfileModal'
 import TweetModal from './TweetModal'
-import ReplyModal from './EditProfileModal/ReplyModal'
+import ReplyModal from './ReplyModal'
 export { EditProfileModal, TweetModal, ReplyModal }
