+++ src/components/share/ReplyTargetText.jsx
@@ -1,8 +1,8 @@
 import styles from 'assets/styles/components/share/texts.module.scss'
 
-const ReplyTargetText = ({ name }) => (
+const ReplyTargetText = ({ name, preText = '回覆' }) => (
   <span className={styles.replyTargetText}>
-    回覆 <span>@{name}</span>
+    {preText} <span>@{name}</span>
   </span>
 )
